backend/requirements.txt
fastapi==0.79.1
uvicorn==0.18.2
mangum==0.15.1
snowflake-connector-python==2.7.11
